[DXP-1322] Update released version
