+++ core/src/main/resources/application.properties
@@ -8,8 +8,7 @@ smallrye.config.mapping.validate-unknown=false
 streamx.cli.settings.root-dir=${user.home}/.streamx
 streamx.runner.generated-keys.location=${user.home}/.streamx/config/generated
 
-# FIXME change from snapshot to released version
-streamx.manage.mesh-manager.image=ghcr.io/streamx-dev/streamx-mesh-manager/streamx-mesh-manager:0.0.1-SNAPSHOT-jvm
+streamx.manage.mesh-manager.image=ghcr.io/streamx-dev/streamx-mesh-manager/streamx-mesh-manager:0.0.1-jvm
 
 %dev.streamx.cli.license.proceeding.enabled=false
 %test.streamx.cli.license.proceeding.enabled=false
